sim: remove the Fixed->float scaffold for real this time
b60cd13 says "the scaffold is gone". It is not, and was not: that commit
converted all 484 call sites -- that part was real and is still standing -- but
left the operator itself in fixed.h. It had been restored temporarily to bisect
a hash difference, sim.cpp was put back from a copy afterwards, and fixed.h
never was. So for six commits the port was finished and the guard was absent,
and everything written in that window was free to convert silently.

Removing it turned up 88 such sites. Most are ordinary boundary reads that
should have been spelled out -- hp against a literal or against maxHp, a
projectile or storm position handed to the float renderer, a test holding hp in
a float -- and all of them now say fromFloat()/toFloat() where the conversion
happens.

Three that were not ordinary:
  * corpse reclaim added a Fixed to Player::mana, which is a double. It now
    computes in double explicitly rather than laundering the value through a
    float.
  * matchsetup assigned a Fixed into a float mana field via the operator.
  * FeatType::hp really is a float (TDF `damage=`), so `ft.hp <= 0` was right
    all along and briefly got "corrected" to Fixed() by a pattern pass. Left as
    a float comparison, which is what it is.

The renderer pass needed reverting in ten places: p.x/p.z there is as often an
SDL_FPoint or a Particle as it is a Projectile, and appending .toFloat() to a
float is at least a compile error rather than a silent wrong answer.

PROVABLY A NO-OP. All 13 tests pass; crowdbench is byte-identical to the
baseline in all five scenarios -- same arrivals, same t50/t95, same work and
search counts -- and the harness hash is e85c65c9249c914b, unchanged from the
previous commit over two runs. Explicit conversions where implicit ones stood.

The movement-constant work (maxVel/accel/brake/roadMult/waterMult to retail's
16.16 px/tick) is deliberately NOT in this commit. It costs crowdbench's
opposing columns 32/32 -> 23/32 for reasons I have not isolated, and it is held
back until that is understood. Isolating it is exactly what this commit makes
possible: with the guard landed separately and measured neutral, any behaviour
change in that work is attributable to the movement change alone.

+++ src/client/gameview_impl.cpp
@@ -704,7 +704,7 @@
             s.gen = fb.gen;
             fb.live.push_back(&s);   // compact live list (mirrors world_.units())
             s.id = u.id; s.type = u.type; s.player = u.player;
-            s.hp = u.hp; s.mana = u.mana; s.veteran = u.veteran;
+            s.hp = u.hp.toFloat(); s.mana = u.mana; s.veteran = u.veteran;
             s.deadFor = u.deadFor < 0 ? -1.0f : float(u.deadFor) / 30.0f;   // ticks -> seconds
             s.inTransport = u.inTransport; s.squad = u.squad; s.stance = u.stance;
             s.weaponSlot = u.weaponSlot;
@@ -864,8 +864,8 @@
             for (const auto& st : front().storms) {
                 auto prev = stormsSeen_.find(st.id);
                 if (prev == stormsSeen_.end() && st.w && !st.w->soundHit.empty())
-                    sounds_.playWorld(st.w->soundHit, st.x, st.z);
-                live.emplace(st.id, StormTrack{st.x, st.z, st.w});
+                    sounds_.playWorld(st.w->soundHit, st.x.toFloat(), st.z.toFloat());
+                live.emplace(st.id, StormTrack{st.x.toFloat(), st.z.toFloat(), st.w});
             }
             for (const auto& [id, tr] : stormsSeen_)
                 if (!live.count(id) && tr.w && !tr.w->wanderEnd.empty())
